Add files via upload
The activity with a counter has the launch mode set to singleInstance which will create only one instance of that activity and redirect all intents for that activity to that task. The retain state attribute will keep the state same as the last time, preventing the activity from resetting. The counter will count the number of time activity b was launched.